feat: use ModalBottomSheet for box creation
